feat(backend): refactor webapp/jiaapi_mock
