+++ development/docker-compose-dev.yml
@@ -53,7 +53,7 @@ services:
     image: golang:1.16.5-buster
     restart: always
     working_dir: /jiaapi
-    command: ["go", "run", "./cmd/standalone/"]
+    command: ["go", "run", "."]
     ports:
       - "5000:5000"
     volumes:
