--- webapp/jiaapi_mock/cmd/standalone/auth.go
@@ -1,38 +1,0 @@
-package main
-
-import (
-	"crypto/ecdsa"
-	"log"
-	"time"
-	_ "embed"
-
-	"github.com/dgrijalva/jwt-go"
-)
-
-const (
-	// lifetime は jwt の発行から失効までの期間を表す。
-	lifetime = 30 * time.Minute
-)
-
-//go:embed ec256-private.pem
-var privateKey []byte
-var jwtSecretKey *ecdsa.PrivateKey
-
-func init() {
-	var err error
-	jwtSecretKey, err = jwt.ParseECPrivateKeyFromPEM(privateKey)
-	if err != nil {
-		log.Fatalf("Unable to parse ECDSA private key: %v", err)
-	}
-}
-
-// Generate は認証に利用する JWT トークンを生成して返す。
-func generateJWT(userID string, now time.Time) (string, error) {
-	token := jwt.NewWithClaims(jwt.SigningMethodES256, jwt.MapClaims{
-		"jia_user_id": userID,
-		"iat":         now.Unix(),
-		"exp":         now.Add(lifetime).Unix(),
-	})
-
-	return token.SignedString(jwtSecretKey)
-}
