+++ webapp/jiaapi_mock/model/utils.go
@@ -1,0 +1,7 @@
+package model
+
+import "strconv"
+
+func getKey(targetIP string, targetPort int, isuUUID string) string {
+	return isuUUID + targetIP + strconv.Itoa(targetPort)
+}
